refactor(multipooler): fold ExpireBackups and VerifyBackups into rpc_backup.go
Co-locate the thin ExpireBackups and VerifyBackups handlers with the
other backup-lifecycle RPCs that delegate to the same backup engine,
matching the one-file-per-area layout used by rpc_consensus.go. Their
tests move alongside the existing backup handler tests.
